Move presentAssistantMessage into its own module
